fix(user): lock tenant OAuth account binding

+++ crystal-resource/src/main/kotlin/com/lovelycatv/crystalframework/resource/service/api/impl/OSSFileResourceServiceImpl.kt
@@ -162,7 +162,7 @@ class OSSFileResourceServiceImpl(
                 PutObjectRequest.newBuilder()
                     .bucket(bucketName)
                     .key(objectKey)
-                    .contentLength(fileLength)
+                    .contentLength(fileLength.toInt())
                     .contentType(fileContentType)
                     .body(BinaryData.fromStream(inputStream))
                     .build()
